Fix Host migration
When receiving a network error (signal for host disconnected), we only handled "Connection Lost" as candidate to initiate a new host.
However, we also receive a "FailureReceived" first, and this then sends all the clients directly back to the lobby, without checking for a new host.

Adds the functionality to add the player over the GameMode-PostLogin